test(infrastructure): move test files into modules/infrastructure/__tests__/ (#708)
Move 5 service tests + 1 route test (renamed to edge-jobs-route.test.ts) into
the infrastructure module. Delete duplicate edge-tunnel-retry.test.ts (covered
by edge-log-fetcher.test.ts). Update vi.mock paths in 4 non-moved consumer tests
(container-logs, search, security-regression, setup) to match new module location.
