+++ backend/src/modules/infrastructure/__tests__/edge-async-log-fetcher.test.ts
@@ -1,13 +1,13 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
-import * as portainer from '../core/portainer/portainer-client.js';
+import * as portainer from '../../../core/portainer/portainer-client.js';
 import {
   initiateEdgeAsyncLogCollection,
   checkEdgeJobStatus,
   retrieveEdgeJobLogs,
   cleanupEdgeJob,
   getEdgeAsyncContainerLogs,
-} from './edge-async-log-fetcher.js';
+} from '../services/edge-async-log-fetcher.js';
 
 describe('edge-async-log-fetcher', () => {
   beforeEach(() => {
